amcl: add global_alt_frame_id to use w/ tf_reverse
When using tf_reverse it is quite handy to be able to compare the
running localizer against another localizer. Add ability to rewrite
the global frame into an alternate frame to make visualization easy.
Also accept initialpose in the alt frame as well via rewrite.

This could be done within the launch file with topic re-writes but
that is expensive when this is a common use case.