amcl: add config. thresh. for uniform poses
For recovery or global localization, allow a weight threshold on the
random pose based on the last laser scan. By default, this weight is
zero (disabled).

The match is controlled by two parameters,
recovery_pose_starting_weight_threshold, which is the first target
weight threshold used when picking a random pose. If the random pose is
not at or above the threshold, the threshold is adjusted down by the
recovery_pose_deweight_multiplier and another pose is checked. This is
repeated until an acceptable pose is found.

Using too high a weight or too high a deweight multiplier will result in
recovery and global localization taking a very long time. Also, the
correct weight to use depends on the sensor model and its configuration.