Improve function pointer implementation
Improving implementation of function pointers.
Function pointers were implemented with static functions and passing an instance of "this" to
create a sudo member function. With C++ we can bind the object "this" to a member function and
pass that as a function pointer. We also don't need to forward declare the function pointer
in the sensor classes.

Removing virtual from child class function prototype.
According to the C++ Style Guide child class function prototypes should not
be declared virtual.