Refactor messy functions
We want to have clean, readable functions so that future writers can understand and develop
the existing code. This refactoring primarily focuses on functions that are too long and have
too many return statements. A function can have many return statements as long as the function
is short. A function can be longer if it is still working on a single, simple goal that should
not be broken into different tasks.