Calculate scanner to footprint tf on each scan
Instead of calculating the transform from the scanner to the footprint
once and using it for all transforms and from all scanners, we should
be calculating that transform every scan and for every scanner separately.
This will allow us to abstract to multiple scanners as well as for moving
scanners or scanners with updated transforms.