Refactor multiple calls to transformPointCloud
Instead of transforming the point cloud multiple times, it is better to multiply
the transforms together and then transform the point cloud by that single
transform. It is also impossible for the pcl_ros::transformPointCloud function
to throw an exception when given a transform so the try-catch should be removed.
Therefore, the function should also be void as it will not fail.