Convert octomap LUT from hash table to vectors
The hash table is not very scalable because it takes too much memory.
It takes less memory  to store the distance itself in uint8 size (as
a ratio to the max distance). Additionally, to save memory we can only
store the values for a column if any z value of that column is anything
other than max occ dist. While this may not reduce the number of
voxels stored as much as the hashmap does, it probably reduces the majority.

Adding isVoxelValid to validate 3D input.