save pose to disk on timer
Save the pose to disk on its own timer. IO can be slow and block the
localization process an indefinite amount of time. To prevent IO from
slowing down localization, save the pose on its own timer. This works
because badger_amcl runs all callbacks in a a multi-threaded spinner.