publish initial pose as TF when no sensor data
If no sensor data is received, but there is a map and odometry, go ahead
and begin publishing the map to odom transform using the initial pose.
Too many other systems depend on the map to odom transform being present
that not having one floods the system log with all kinds of transform
timeouts, muddying the true issue (that no sensor data was being
received). There may even be times when the sensors are intentionally
off when power is being saved, and if somehow AMCL gets restarted in
that state, it should continue to publish the initial pose as the
current map to odom TF.

As a part of this change, consolidate the identical updatePose methods
in Node2D and Node3D into the Node itself to keep from having to
implement this functionality in two places. The old Node updatePose did
not return a boolean, and therefore the old Node2D/Node3D did not check
if they worked. Since updatePose can now be called prior to the particle
filter being setup (and fails if so), it must be made to return a
boolean. Instead of implementing this change in both Node2D and Node3D,
just consolidate and remove their updatePose methods since they were
identical anyway.