publish initial pose as TF when no sensor data
If no sensor data is received, but there is a map and odometry, go ahead
and begin publishing the map to odom transform using the initial pose.
Too many other systems depend on the map to odom transform being present
that not having one floods the system log with all kinds of transform
timeouts, muddying the true issue (that no sensor data was being
received). There may even be times when the sensors are intentionally
off when power is being saved, and if somehow AMCL gets restarted in
that state, it should continue to publish the initial pose as the
current map to odom TF.