ignore odom absolute motion when too big
If there has been too much absolute motion since the last update,
ignore it and use the simple pose delta for the motion estimate.
If the absolute motion were used, too much noise would be added.
This should only happen if the sensor data is coming in so slowly that
enough absolute motion could accumulate to be twice the normal
threshold, as otherwise the filter would have already been updated.