Fix bug related to multiple maps
The function updateDistancesLUT resets the octree_ pointer. When
the updateDistancesLUT function was run and then a new 2D map was
received before receiving a new Octomap message, the octree_ pointer
would still be null resulting in an exception. This change ensures that
a new octomap has been received before updating the LUT.