Fix bug that would score all randomly generated poses as perfect
(No functional change when `recovery_alpha_slow` and
`recovery_alpha_fast` are both set to 0 because in that case 0 poses are
generated randomly.)

When a vector is initialized with a structure in curly braces, what is
in the vector is a _copy_ of the structure, not the structure itself.
Therefore, in `scorePose()` the weight that was being returned was the
original structure for which the weight had been set to 1.0, not the
calculated weight (which was stored in the copy).

Since `scorePose()` is called by `uniformPoseGenerator()` to assess the
quality of a randomly generated pose, poses which poorly matched the
observation were not being discarded as they should have been leading to
a pollution of the particle cloud which was then getting averaged
together (with good poses) to come up with the next pose estimate which
incorrectly skewed the pose.

ROS-1692: Robot appears to drift during Shelf Nav