Rename convertWorldToMap and convertMapToWorld
The map functions convertWorldToMap and convertMapToWorld are confusing
as they create a new concept of "world" and change the concept of "map".
These functions are really rasterizing a point in the map frame to
indices and unrasterizing indices to a point in the map frame. The
ranamed functions are rasterize and unrasterize which better reflect
their goals.