Publish pose before resampling particle cloud
This changes a key design element of the original amcl upstream package.
The design choice they made was to first resample the particle cloud
and then calculate and publish the pose estimate. This approach removes
all knowledge of the particle weights previously acquired as the
particle weights are normalized after resampling. The problem with this
is that any bias that affects the particle filter will then affect the
pose estimation. The new approach is to first publish the pose
estimation and then resample the particles. This allows the pose
estimation to be made based on the weights of the particles that have
accumulated since the last resampling.